Split collections and item classes into separate files and out of base.py
The new attribute system will involve meta classes other helper objects
I prefer to keep separate from the objects that use them for neatness.
